Fixups found during work on release testing
changelog: need to have a recent version to match the testing of the release.

debian/rules: project-builder overwrites the debian directory; move fix-install.sh into dist.
  Update the ChangeLog name to match what project-builder wants and remove the .mtn since we're
  now on git.

fix-install.sh: remove build user's path from built file.

make-release-changelog.sh: Update the changelog generation to work with git rather than monotone.

dist/rules: Add more to the help.  Update the way we do prep-release so that we can make it a
  dependency for snapshot-tar.  Add rule to generate a snapshot of the source code so we can test
  multiple things without having to commit each time.

Lintel.spec.in: same fix as changelog

make-tar.sh: Update to work with git, to verify that the necessary files are generated, and fix
  some bugs that occur from running the script multiple times or not having the SOURCES directory.

+++ redhat/make-tar.sh
@@ -1,4 +1,5 @@
 #!/bin/sh
+set -e -x
 case `pwd` in
    */Lintel | */Lintel.0.20[0-9][0-9].[0-9][0-9].[0-9][0-9])
       : 
@@ -8,14 +9,23 @@ case `pwd` in
       ;;
 esac
 
-./dist/make-release-changelog.sh
+if [ -d .git ]; then
+    ./dist/make-release-changelog.sh
+fi
+
+if [ ! -f Release.info -o ! -f ChangeLog ]; then
+    echo "Error: Ought to either both Release.info and Changelog.mtn from a release or should have just created it from monotone repository"
+    exit 1
+fi
+
 . redhat/get-version.sh
 
 cwd=`pwd`
 dir=`basename $cwd`
 cd ..
-[ "$dir" == "Lintel-$VERSION" ] || ln -s $dir Lintel-$VERSION
-tar cvvfhz Lintel-$VERSION.tar.gz --exclude=Lintel-$VERSION/_MTN Lintel-$VERSION/
-[ "$dir" == "Lintel-$VERSION" ] || rm Lintel-$VERSION
+[ "$dir" = "Lintel-$VERSION" -o -d Lintel-$VERSION ] || ln -s $dir Lintel-$VERSION
+tar cvvfhz Lintel-$VERSION.tar.gz --exclude=Lintel-$VERSION/.git Lintel-$VERSION/
+[ "$dir" = "Lintel-$VERSION" ] || rm Lintel-$VERSION
 
+mkdir -p $rpm_topdir/SOURCES
 cp Lintel-$VERSION.tar.gz $rpm_topdir/SOURCES/Lintel-$VERSION.tar.gz
